Added second browser

+++ main.py
@@ -10,19 +10,19 @@ import pickle
 import os
 
 
-def do_looting():
-    balance = float(driver.find_element(By.CSS_SELECTOR,BALANCE_ID).text[-5])
+def do_looting(driver, ITEMS):
+    balance = float(driver.find_element(By.CSS_SELECTOR,BALANCE_ID).text[:-1].replace(',', '.'))
     print(balance)
     driver.find_element(By.XPATH,SKIN_DATA_SWITCH).click()
     time.sleep(1)
 
     while balance > 1:
+        print('Cycle')
         for item in ITEMS:
-            get_items(item['link'], item['desired_float'])
-            time.sleep(5)
+            get_items(driver, item['link'], item['desired_float'])
 
 
-def get_items(itemLink, desiredFloat):
+def get_items(driver, itemLink, desiredFloat):
     pageNumbers = ['0', '100', '200', '300']
 
 
@@ -41,8 +41,8 @@ def get_items(itemLink, desiredFloat):
             except:
                 print('Button not found')
                 timesButtonNotFound += 1
-                if timesButtonNotFound >= 10:
-                    get_items(itemLink,desiredFloat)
+                if timesButtonNotFound >= 5:
+                    get_items(driver, itemLink, desiredFloat)
                     print('Reloading page')
                     return 0
 
@@ -55,9 +55,9 @@ def get_items(itemLink, desiredFloat):
             if itemFloat < desiredFloat:
                 floatValues.append(itemFloat)
                 print(itemFloat)
-                buy_item(item, itemLink, desiredFloat)
+                buy_item(driver, item, itemLink, desiredFloat)
 
-def buy_item(item, itemLink, desiredFloat):
+def buy_item(driver, item, itemLink, desiredFloat):
     try:
         item.find_element(By.XPATH, '../../../../div[2]/div[1]/div/a').click()
         time.sleep(3)
@@ -66,12 +66,12 @@ def buy_item(item, itemLink, desiredFloat):
         driver.find_element(By.XPATH, '//*[@id="market_buynow_dialog_purchase"]').click()
         time.sleep(2)
         driver.find_element(By.XPATH, '//*[@id="market_buynow_dialog_close"]').click()
-        time.sleep(1)
+        time.sleep(0.5)
     except:
         messagebox("Buy failed")
-        get_items(itemLink, desiredFloat)
+        get_items(driver, itemLink, desiredFloat)
 
-def save_session():
+def save_session(driver):
     try:
         with open('session.pkl', 'wb') as session_file:
             pickle.dump(driver.get_cookies(), session_file)
@@ -79,7 +79,7 @@ def save_session():
 
     except:
         print('Session save failed')
-def load_session():
+def load_session(driver):
     time.sleep(0.3)
     driver.delete_all_cookies()
 
@@ -94,11 +94,14 @@ def load_session():
         print('cookies not found')
         return  0
 
-    print(cookies)
+    if len(cookies) == 0:
+        print('Cookies are empty')
+        return 0
+
     if cookies[0]['expiry'] <= int(time.time()):
         print(int(time.time()))
         os.remove('session.pkl')
-        save_session()
+        save_session(driver)
         with open('session.pkl', 'rb') as session_file:
             print('opened file')
             cookies = pickle.load(session_file)
@@ -113,25 +116,24 @@ def load_session():
     except:
         print('cookies not added')
         pass
-def start_button_click():
-    do_looting()
-    messagebox.showinfo("Info", "Start button clicked")
+# def start_button_click():
+#     do_looting()
+#     messagebox.showinfo("Info", "Start button clicked")
+#
+# def stop_button_click():
+#     messagebox.showinfo("Info", "Stop button clicked")
+#
+# def exit_button_click():
+#     global overlay_window
+#     if messagebox.askyesno("Exit", "Do you want to save cookies?"):
+#         save_session()
+#     else:
+#         os.remove('session.pkl')
+#
+#     driver.close()
+#     exit()
 
-def stop_button_click():
-    messagebox.showinfo("Info", "Stop button clicked")
-
-def exit_button_click():
-    global overlay_window
-    if messagebox.askyesno("Exit", "Do you want to save cookies?"):
-        save_session()
-    else:
-        os.remove('session.pkl')
-
-    driver.close()
-    exit()
-
-
-def check_login():
+def check_login(driver):
     try:
         driver.find_element(By.XPATH, LOGIN_BUTTON)
         print('Not logged in')
@@ -141,22 +143,41 @@ def check_login():
         return True
 
 
+def lunch(driver, ITEMS):
+    driver.get(STEAM_LINK)
+    load_session(driver)
+    driver.get(STEAM_LINK)
+    print(driver.get_cookies())
+    time.sleep(1)
+
+    if check_login(driver):
+        driver.get(ITEMS[0]['link'])
+        do_looting(driver, ITEMS)
+    else:
+        if messagebox.askyesno('Check log in', 'Have you logged in the account?'):
+            save_session(driver)
+            driver.get(ITEMS[0]['link'])
+            do_looting(driver, ITEMS)
+
 #----------------------------------------- MAIN BODY ------------------------------------------------#
 
 #ITEMS#
 ITEMS = [
 #Fracture STATTREK
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
-         {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MAG-7%20%7C%20Monster%20Call%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MAC-10%20%7C%20Allure%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20Tec-9%20%7C%20Brother%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20Galil%20AR%20%7C%20Connexion%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
+         # {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         # {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MAG-7%20%7C%20Monster%20Call%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         # {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MAC-10%20%7C%20Allure%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         # {'link':'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20Tec-9%20%7C%20Brother%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         # {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20Galil%20AR%20%7C%20Connexion%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
 #Fracture
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199},
-         {'link': 'https://steamcommunity.com/market/listings/730/StatTrak%E2%84%A2%20MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=0&count=100', 'desired_float': 0.199}]
+         {'link': 'https://steamcommunity.com/market/listings/730/Galil%20AR%20%7C%20Connexion%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         {'link': 'https://steamcommunity.com/market/listings/730/MAG-7%20%7C%20Monster%20Call%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         {'link': 'https://steamcommunity.com/market/listings/730/MP5-SD%20%7C%20Kitbash%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         {'link': 'https://steamcommunity.com/market/listings/730/Tec-9%20%7C%20Brother%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199},
+         {'link': 'https://steamcommunity.com/market/listings/730/MAC-10%20%7C%20Allure%20%28Field-Tested%29?query=&start=', 'desired_float': 0.199}]
+
+ITEMS_SMALL = [ITEMS[0], ITEMS[1]]
+ITEMS_BIG = [ITEMS[2], ITEMS[3], ITEMS[4]]
 
 
 #CONSTANTS#
@@ -175,38 +196,30 @@ chrome_options.add_experimental_option('detach', True)
 chrome_options.add_argument('--ignore-certificate-error')
 chrome_options.add_argument('--ignore-ssl-errors')
 chrome_options.add_extension('1.18.41_0.crx')
+
 driver = webdriver.Chrome(options=chrome_options)
 driver.maximize_window()
 
-driver.get(STEAM_LINK)
-load_session()
-driver.get(STEAM_LINK)
-print(driver.get_cookies())
-time.sleep(1)
+driver2 = webdriver.Chrome(options=chrome_options)
+driver2.maximize_window()
 
-if check_login():
-    driver.get(ITEMS[0]['link'])
-    do_looting()
-else:
-    if messagebox.askyesno('Check log in', 'Have you logged in the account?'):
-        save_session()
-        driver.get(ITEMS[0]['link'])
-        do_looting()
+thread1 = Thread(target=lunch, args=[driver, ITEMS_BIG])
+thread1.start()
+thread2 = Thread(target=lunch, args=[driver2, ITEMS_SMALL])
+thread2.start()
 
-driver.get(ITEMS[0]['link'])
-do_looting()
 
-overlay_window = tk.Tk()
-overlay_window.title("Modern Buttons Window")
-style = ttk.Style()
-style.configure("TButton", padding=10, font=("Helvetica", 12))
-
-start_button = ttk.Button(overlay_window, text="Start", command=start_button_click)
-stop_button = ttk.Button(overlay_window, text="Stop", command=stop_button_click)
-exit_button = ttk.Button(overlay_window, text="Exit", command=exit_button_click)
-
-start_button.pack(pady=10)
-stop_button.pack(pady=10)
-exit_button.pack(pady=10)
-
-overlay_window.mainloop()
+# overlay_window = tk.Tk()
+# overlay_window.title("Modern Buttons Window")
+# style = ttk.Style()
+# style.configure("TButton", padding=10, font=("Helvetica", 12))
+#
+# start_button = ttk.Button(overlay_window, text="Start", command=start_button_click)
+# stop_button = ttk.Button(overlay_window, text="Stop", command=stop_button_click)
+# exit_button = ttk.Button(overlay_window, text="Exit", command=exit_button_click)
+#
+# start_button.pack(pady=10)
+# stop_button.pack(pady=10)
+# exit_button.pack(pady=10)
+#
+# overlay_window.mainloop()
